--- react/src/pages/LoginPage.jsx
@@ -1,8 +1,0 @@
-
-export default function LoginPage(){
-    return (
-        <>
-            <h1>Login Page</h1>
-        </>
-    );
-}
