--- react/src/pages/RegisterPage.jsx
@@ -1,8 +1,0 @@
-
-export default function RegisterPage(){
-    return (
-        <>
-            <h1>Register Page</h1>
-        </>
-    );
-}
